More (very ugly) station/module merging code.
The earlier code didn't modify the hardware IDs on the modules and that ended up
creating an even bigger mess. This code works to preserve a single module with
the correct hardware id and deletes the others, including their configurations
and provisions while also merging the data. It's messy as hell.