Module ids are mapped based on last sensor reading.
When we added associated stations and the ability to combine/merge
sensors for them, this code to remap module IDs back to the original
station was added. This code breaks when a module moves from one station
to another and both stations are present in data viz.

So, as a quick solution we sort by the sensor reading date. I
believe this can and should be improved further. For example,
the instances where this remapping happens could probably be narrowed
significantly.