Filter scrub events based on initial mousedown.
The earlier fix for this caused the #13, where the drop downs would
close immediately, and the fix for _that_ regressed us to only updating
the displayed time on a scrub in a few situations.

What was happening was vega emits the brush signal quite frequently, not
just when the user is actively scrubbed. Then, mouseup would occur when
clicking drop downs and then it would emit the scrub/time-zoom event and
so things would refresh and close the drop down. The fix would filter
the mouseup's to only occur over the vega viz, which isn't what we want.

The new fix remembers if we're scrubbing by watching mousedown and
remembering and will only emit scrub/time-zoom events if that is true. I
_think_ and hope that will be enough.