Only create configurations for schema mappings w/o bays.
This is kind of a hack but it'll work for now to get LoRa working. Before this,
we were consistently creating and overriding the visible station configuration for
FieldKit stations, which would cause us to "lose" the configuration that actually
matched the station, with the real modules, etc... This will just query
if all the module schemas give bay numbers.