feat: Optimistic locking on dirty sensor data rows.
This will verify we delete the same number of rows we queried for, so
that processing two messages concurrently will cause one to fail.
Because jobs are also in the database this will cause the published
refreshes to roll back on the failed job as well.

Another thing we could have done is to delete dirty rows within the time
range we published.