fix: Column getters in CSV export are per (metaId, moduleId)
Prior to this they were simply tied to module Ids, but this won't work
if a module is moved from one bay to another and we try and export the
station's data. This is because we'll have two candidate fieldSets, one
for each bay position, and w/o the constraint we end up pulling the
sensor data for whatever module is occupying that bay after the move.

This took some chasing, so there may be a little in here that's
superfluous, but it felt worth keeping. The key here is the extra
condition, checkingg for the same metaId on the fieldset being used and
the change that causes us to create a fieldSet for each (metaId,
moduleId) combination.